Removing RestClient and Apache HTTP Client in favor of Net::Http
